# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "489c38c55debbddbfae2b1d58cab5257",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais invalidas.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user and delete JWT token",
                "operationId": "f7d09b5a0aa710a48ddf3eb44274a8bd",
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "474f59d33ec7ae95108f77316982c3aa",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "E-mail já está sendo utilizado.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "409": {
                        "description": "Erro ao criar usuário.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/schedule": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "List all schedulings for the authenticated user.",
                "operationId": "770209026e4bbff78b8f3abeffa313f9",
                "responses": {
                    "200": {
                        "description": "Listagem de agendamentos realizada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Register a Schedule.",
                "operationId": "eedf19fd09e11816cd56ca44811968b4",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "query",
                        "description": "Type's ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Schedule's title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Schedule's description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "starting_at",
                        "in": "query",
                        "description": "Starting Date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "due_at",
                        "in": "query",
                        "description": "Due Date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "conclusion_at",
                        "in": "query",
                        "description": "Conclusion Date",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agendamento atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Register a type.",
                "operationId": "23a3ec1c1d504d256eef0f9facfacb44",
                "parameters": [
                    {
                        "name": "type_id",
                        "in": "query",
                        "description": "Type's ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Schedule's title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Schedule's description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "starting_at",
                        "in": "query",
                        "description": "Starting Date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "due_at",
                        "in": "query",
                        "description": "Due Date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "conclusion_at",
                        "in": "query",
                        "description": "Conclusion Date",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agendamento registrado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/schedule/{id}": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Show a specific scheduling for the authenticated user.",
                "operationId": "8c7f0c61e1c3822598cede1d9f72b3cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Type's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de gendamento recuperado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Delete a Schedule's register.",
                "operationId": "7691ba4f40cad0c8de2945100156e3c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Type's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agendamento excluído com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/type": {
            "get": {
                "tags": [
                    "Type"
                ],
                "summary": "List all types of scheduling for the authenticated user.",
                "operationId": "5ecf20319530ee059d73107557fd14a4",
                "responses": {
                    "200": {
                        "description": "Listagem de tipo de gendamento relizada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Type"
                ],
                "summary": "Register a type.",
                "operationId": "d50e7c4c2e30cf78915446eb1edce1ec",
                "parameters": [
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Type's description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de gendamento registrado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/type/{id}": {
            "get": {
                "tags": [
                    "Type"
                ],
                "summary": "Show a specific type of scheduling for the authenticated user.",
                "operationId": "21e37d7490a584a32693d79fbe0cf466",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Type's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de gendamento recuperado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Type"
                ],
                "summary": "Edit a Type's register.",
                "operationId": "f446c9d715b03de657274fef7603fdd4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Type's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de gendamento atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Type"
                ],
                "summary": "Delete a Type's register.",
                "operationId": "c62846a951e9a0bf61a91113a9cb97ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Type's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de gendamento excluído com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}